The implementation of KafkaWriteStream caches a Vert.x context to interact with the Kafka producer. This same context is used in order to serialize producer interactions.
It is desirable to rely instead on the caller Vert.x context, so that producer interactions uses the caller context instead of the cached context.

The KafkaWriteStream implementation changes now to use the caller Vert.x context and provide a task queue to serialize producer interactions instead of relying on the internal task queue of the previously cached context.