Fix ProducerMockTest#testWriteWithSimulatedError
The test was implemented on the assumption of a "sticky" context (always the same context returned when getOrCreateContext is invoked on a non-Vert.x thread).

But in Vert.x 5, sticky contexts are gone. So in order to catch the exception, the handler must be set on the Vert.x instance.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>